build: move all shareable components to packages
