+++ apps/Tableau/components/body/components/side/components/boardDelete.tsx
@@ -2,8 +2,7 @@ import React, { useCallback } from 'react'
 import axios from 'axios'
 import { IconButton, Tooltip, useDisclosure } from '@chakra-ui/react'
 import { BsTrashFill } from 'react-icons/bs'
-
-import { DeleteModal } from '../../modal/deleteModal'
+import { DeleteModal } from 'shared-components'
 import { useTableauMutation } from 'shared-hooks'
 import { IBoardWithAllRelation } from '../../../../../types/types'
 import { useAtom } from 'jotai'
