+++ apps/Tableau/components/body/components/side/components/boardEditForm.tsx
@@ -1,7 +1,7 @@
 import { Formik, FormikHelpers } from 'formik'
 import React, { useCallback, useMemo } from 'react'
 import * as Yup from 'yup'
-import { TextInput } from '../../../../textInput'
+import { TextInput } from 'shared-components'
 import { Button, ButtonGroup, Flex, Text } from '@chakra-ui/react'
 import { WarningIcon } from '@chakra-ui/icons'
 import { IBoardWithAllRelation } from '../../../../../types/types'
