+++ apps/Tableau/package.json
@@ -50,7 +50,8 @@
     "yup": "^0.32.11",
     "shared-atoms": "*",
     "shared-utils": "*",
-    "shared-hooks": "*"
+    "shared-hooks": "*",
+    "shared-components": "*"
   },
   "devDependencies": {
     "@types/node": "^18.11.18",
