+++ packages/components/package.json
@@ -2,10 +2,49 @@
   "name": "shared-components",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
-  "license": "ISC"
+  "dependencies": {
+    "@chakra-ui/icons": "^2.0.16",
+    "@chakra-ui/react": "^2.4.6",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fontsource/open-sans": "^4.5.13",
+    "@fontsource/raleway": "^4.5.11",
+    "@next-auth/prisma-adapter": "^1.0.5",
+    "@next/bundle-analyzer": "^13.1.6",
+    "@next/font": "13.1.1",
+    "@prisma/client": "^4.8.0",
+    "@tanstack/react-query": "^4.22.0",
+    "@types/react": "18.0.26",
+    "@types/react-dom": "18.0.10",
+    "axios": "^1.2.3",
+    "dotenv-cli": "^6.0.0",
+    "eslint": "8.31.0",
+    "eslint-config-next": "13.1.1",
+    "formik": "^2.2.9",
+    "framer-motion": "^6.5.1",
+    "jotai": "^1.13.1",
+    "next": "13.1.1",
+    "next-auth": "^4.18.7",
+    "react": "18.2.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
+    "react-dom": "18.2.0",
+    "react-icons": "^4.7.1",
+    "yup": "^0.32.11",
+    "shared-atoms": "*",
+    "shared-utils": "*",
+    "shared-hooks": "*"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.18",
+    "prettier": "^2.8.1",
+    "prisma": "^4.8.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
+  }
 }
