+++ packages/components/src/components/columns/columnTask.tsx
@@ -17,7 +17,7 @@ import { useDrop } from 'react-dnd'
 import { TaskItemType } from 'shared-utils'
 import { ITaskEditFormikValues } from './taskEdit'
 import { RefetchBoardAtom } from 'shared-atoms'
-import { IBoardWithAllRelation, IFullStatus } from '../../../../../types/types'
+import { IBoardWithAllRelation, IFullStatus } from '../../types/types'
 import { ColumnTaskMove } from './columnTaskMove'
 import { useTableauMutation } from 'shared-hooks'
 import { ColumnEdit } from './columnEdit'
