+++ packages/components/src/components/columns/columnTaskMove.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useMemo } from 'react'
 import axios from 'axios'
 import { IconButton, Tooltip } from '@chakra-ui/react'
 import { ArrowBackIcon, ArrowForwardIcon } from '@chakra-ui/icons'
-import { IFullStatus } from '../../../../../types/types'
+import { IFullStatus } from '../../types/types'
 import { useAtom } from 'jotai'
 import { BoardAtom, RefetchBoardAtom } from 'shared-atoms'
 import { useTableauMutation, useSwapEntity } from 'shared-hooks'
