+++ packages/components/src/components/columns/columnTaskNewForm.tsx
@@ -3,10 +3,10 @@ import { BoardAtom } from 'shared-atoms'
 import React, { useCallback, useMemo } from 'react'
 import * as Yup from 'yup'
 import { Formik, FormikHelpers } from 'formik'
-import { TextInput } from '../../../../textInput'
+import { TextInput } from '../textInput'
 import { Button, ButtonGroup, Flex, Text } from '@chakra-ui/react'
 import { WarningIcon } from '@chakra-ui/icons'
-import { IFullStatus } from '../../../../../types/types'
+import { IFullStatus } from '../../types/types'
 
 export type ITaskNewFormikValues = {
     name: string
