+++ packages/components/src/components/finder/index.tsx
@@ -9,7 +9,7 @@ import axios, { AxiosResponse } from 'axios'
 import { Box, Flex, Text, useDisclosure } from '@chakra-ui/react'
 import { useTableauMutation } from 'shared-hooks'
 import { Task } from '.prisma/client'
-import { TaskItem } from '../../../body/components/board/components/taskItem'
+import { TaskItem } from '../columns/taskItem'
 import { getScrollbarStyle } from 'shared-utils'
 import { SearchInput } from './searchInput'
 import { RiEmotionSadLine } from 'react-icons/ri'
