+++ packages/components/src/index.ts
@@ -1,0 +1,7 @@
+export { Columns } from './components/columns/columns'
+export { TaskItem } from './components/columns/taskItem'
+export { DeleteModal } from './components/columns/modal/deleteModal'
+export { Finder } from './components/finder'
+export { Profile } from './components/profile'
+export { TextInput } from './components/textInput'
+export { LoadingProvider } from './components/loadingProvider'
