+++ packages/components/src/types/types.ts
@@ -1,0 +1,11 @@
+import { Board, Status, StatusBoard, Task, User } from '.prisma/client'
+
+export type IBoardWithAllRelation = Board & {
+    Task: Task[]
+    Status: IFullStatus[]
+    user: User
+}
+
+export type IFullStatus = StatusBoard & {
+    status: Status
+}
